findonpath

Searches for a file on the system environment path

Usage

findonpath(fn, status = TRUE, quote = TRUE)

Arguments

fn

The filename to search for (without a path)

status

Show messages

quote

Enclose returned filename in quotes

Value

The full path and name of the found file, or NULL if not found

Note

This will return the first found occurence of file fn, searching 1) the current working directory, 2) the user's R 'home' directory, then 3) the directories on the operating system environment 'path'