Cameras

Returns the names of known cameras

Read a cameras.csv file

uas_cameras(names_only = TRUE)

uas_readcameras(cameras_fn)

Arguments

names_only
Return only camera names

cameras_fn
The file name (including path) of a cameras.csv file

Value

A character vector containing the names of cameras known to the uasimg package

Details

In order to estimate the GSD and image footprint, certain properties of the camera are required, including the physical dimension of the sensor. Properties of several commonly used cameras have been collected and are stored in a csv file.

If your drone camera is not in the database, you'll get an error message when you try to run uas_info. In this case, you can contact the package author by email, and/or start an issue on GitHub. Alternately, you can create your own CSV file based on the one that comes with the package (to see where that lives, run system.file("cameras/cameras.csv", package = "uasimg")), and pass the file name as the value of cameras in uas_info.

This utility function will read a cameras.csv file. Note certain columns are required. To see an example of a valid CSV file, run uas_cameras(names_only = FALSE)

Functions

uas_readcameras: Read cameras csv file

See also