Export location info from a UAS image collection to Shapefile

Export geometry(s) of a UAS image collection to Shapefile

uas_exp_shp(
  x,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  shp_dir = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

x
A list of class 'uas_info'

ctr
Export the image centroids as a Shapefile, T/F or a filename

fp
Export the image footprints as a Shapefile, T/F or a filename

mcp
Export the minimum convex polygon of the image footprints as a Shapefile, T/F or a filename

shp_dir
The directory where the Shapefiles should be saved. If NULL, they will be saved in a 'map' sub-directory of the image folder

create_dir
Create the output directory if it doesn't exist, T/F

overwrite
Overwrite existing files, T/F

quiet
Supress messages and printing of the pandoc command line, T/F

Value

A vector of Shapefile filenames

Details

ctr, fp, and mcp can be logical (TRUE/FALSE) or a filename (minus the shp extension). If logical values are passed, a Shapefile name will be constructed based on the name of parent folder. You can specify which directory the Shapefiles will be exported to using the shp_dir argument. The default is to save them in a sub-directory of the images directory called 'map'.

See also