Manage cache directory

View and set the directory where extracted EXIF data are cached

Usage

uas_getcache(default = TRUE, quiet = FALSE)

uas_setcache(dir = "~/.R/uasimg", write = FALSE, quiet = FALSE)

uas_clearcache(quiet = FALSE)

Arguments

default

Use a default cache directory if no other has been set

quiet

Suppress messages

dir

The directory for cached EXIF data (must exist)

write

Write directory location to .Renviron

Details

Extracting exif data from a large number of images can take awhile. To avoid having to do this more than once for the same folder of images, the results can be saved (or cached) to a local directory of your choice. When cache = TRUE in uas_info, R will first look to see if EXIF data has already been extracted for the image folder, and if so use it instead of running exiftool.

Cached results are saved as native R objects. Cache files store the EXIF data for a folder of images, however if images are removed or added from a directory, any cached results will be nullified and exiftool will run again. Cached EXIF data does not include supplemental flight metadata that you provide with metadata text files (see uas_metadata_make, such as the pilot's name or location name.

uas_getcache() retrieves the current cache directory, and uas_setcache() sets it. The default location is (~/.R/uasimg). When uas_setcache() is run with write = TRUE, the setting will be persistent across R sessions (generally recommended). uas_clearcache() will delete cached EXIF data, which is sometimes called for after a package update.

Functions

uas_setcache(): Set cache directory

uas_clearcache(): Clear cache directory

See also