Parse an image collection into individual flights

Parse an image collection into individual flights

Usage

uas_grp_flt(
  x,
  interflt_val = 10,
  interflt_units = c("med_int", "secs")[1],
  init_fltnum = 1,
  min_images = 5,
  cross_dirs = TRUE,
  options = NULL,
  quiet = FALSE
)

Arguments

x

A list of class 'uas_info'

interflt_val

Time value between flights

interflt_units

Time units between flights, see details "med_int" or "secs"

init_fltnum

Initial flight number

min_images

Minimum number of images to be considered a flight

cross_dirs

Parse the images in directories collectively

options

Options and overrides for creating groups

quiet

Suppress messages

Details

cross_dirs means images in all directories will be examined when look for flights. This would be appropriate if the images from one flight were spread across multiple folders.