Move UAS images into sub-directories

Move UAS images into sub-directories

Usage

uas_move(
  x,
  flt = NULL,
  tree,
  outdir_base,
  req_all_fltmdflds = TRUE,
  create_dirs = "ask",
  imgs_action = c("copy", "move", "none")[1],
  imgs_prepend_fn = FALSE,
  write_metadata = TRUE,
  preview_only = FALSE,
  tb_action = imgs_action,
  map_action = c("copy", "move", "none")[3],
  quiet = FALSE,
  flt_idx = deprecated()
)

Arguments

x

A list of class 'uas_info'

flt

Flight(s) in x to process (character or numeric vector, default is all)

tree

Directory tree template filename or character vector, see Details

outdir_base

Output directory root

req_all_fltmdflds

Require all flight metadata fields in the directory tree template to be defined

create_dirs

Create the output directory tree

imgs_action

The action to take with images

imgs_prepend_fn

Whether th prepend image file names with datestamp (to ensure they'll be unique)

write_metadata

Write a metadtata.txt file in the output image folder

preview_only

Preview the directory tree only

tb_action

The action to take with thumbnail images saved in the default location

map_action

The action to take with the contents of the map folder

quiet

Suppress messages

flt_idx

`r lifecycle::badge("deprecated")` Use `flt` instead

Value

A vector of directory locations where images were copied / moved

Details

req_all_fltmdflds means don't move anything unless all uas_info objects have all tokens in the directory tree

See also