Manage supplemental metadata fields

Set and view a default set of supplemental metadata fields

Usage

uas_setflds(flds = NULL, reset = FALSE, quiet = FALSE)

uas_getflds()

uas_flds_oem()

Arguments

flds

Character vector of metadata fields

reset

Use a default set of metadata fields

quiet

Suppress messages

Details

These functions allow you to define a default set of supplemental metadata fields that can be assigned image collections either as a function argument, or a small text file placed in the directory with the images. Note in this context 'supplemental metadata' refers to bits of information that can not be automatically extracted from the images themselves (e.g., the name of the project or pilot.)

Creating a default set of supplemental metadata fields does not make them mandatory, nor does it limit the metadata fields you can use. The default set serves as a template when a new metadata.txt file is created.

Functions

uas_setflds(): Set default fields for image collection metadata.

uas_getflds(): Get default supplemental metadata field names.

uas_flds_oem(): Get a standard set of image collection metedata fields

See also