Create image thumbnails

Move UAS images into sub-directories by group

Usage

uas_thumbnails_make(
  x,
  flt = NULL,
  output_dir = NULL,
  tb_width = 400,
  rotate = FALSE,
  overwrite = FALSE,
  use_magick = FALSE,
  stats = FALSE,
  quiet = FALSE,
  flt_idx = deprecated()
)

Arguments

x

A list of class 'uas_info'

flt

Flight(s) in x to process (character or numeric vector, default is all)

output_dir

Output directory

tb_width

Thumbnail width

rotate

Rotate the thumbnails by the camera yaw, Logical

overwrite

Overwrite existing files

use_magick

Use functions from the magick package

stats

Report the amount of time it takes to create each thumbnail, logical

quiet

Suppress messages

flt_idx

`r lifecycle::badge("deprecated")` Use `flt` instead

Value

A named list (one element for each directory processed) of thumbnail files created in the output directory

Details

This will create thumbnail images for the drone images in x. The default output folder is a sub-directory of each image folder called map/tb, which will be created if needed. This location can be overridden with output_dir. The dimensions of the thumbnails is determined by tb_width, from which the height is set automatically.

flt allows you to specify a subset of image folders in x to process. You can pass a vector of flight names (use names(x) to see what those are) or integers.

rotate will rotate the thumbnails by the camera yaw. This can make it easier to match up ground features when viewing the thumbnails in a flight report. Note when thumbnails are rotated the tb_width parameter sets the width of the image before the rotation. The width and height of the rotated thumbnail will vary according to the angle of rotation. If your rotations look off feel free to contact the package author, as this feature is still experimental (i.e., some drones record the yaw of the drone and the yaw of the gimbal camera separately).

Thumbnail files will be given an 8-character suffix that looks random but is actually generated from the image contents. This is to prevent clashes when thumnbail files from different flights are 'gathered' into a single folder attached to a Table of Contents folder (see uas_toc).

If use_magick = TRUE, it will use resizing functions from the magick package. This is slower than the equivalent functions from the imager package (the default), but may be necessary if you are processing TIFs and don't have ImageMagick installed on your computer (which imager requires to read TIFs).

See also