Export flight area and camera locations for GIS

Export geometry(s) from a flight to KML and Shapefile

uas_exp_kml(
  x,
  img_dir = NULL,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  combine_feats = FALSE,
  combine_fn = NULL,
  output_dir = NULL,
  out_fnbase = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)

uas_exp_shp(
  x,
  img_dir = NULL,
  ctr = FALSE,
  fp = FALSE,
  mcp = FALSE,
  combine_feats = FALSE,
  combine_fn = NULL,
  output_dir = NULL,
  out_fnbase = NULL,
  create_dir = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)

Arguments

x
A list of class 'uas_info'

img_dir
Image directories in x to process, character

ctr
Export the image centroids, Logical

fp
Export the image footprints, Logical

mcp
Export the minimum convex polygon of the image footprints, logical

combine_feats
Combine features into a single Shapefile / KML, logical

combine_fn
File name (minus path and extension) for the combined features layer

output_dir
The output directory. If NULL, the layers will be saved in a 'map' sub-directory of the image folder

out_fnbase
The base of output filenames

create_dir
Create the output directory if it doesn't exist, logical

overwrite
Overwrite existing files, Logical

quiet
Suppress messages, Logical

Value

A vector of filenames

Details

img_dir allows you to specify a subset of image folders in x to process (use names(x) to see what those are).

ctr, fp, and mcp (all TRUE/FALSE) specify which geometry(s) to export. Output file names will be generated from the flight metadata saved in the uas_info object. Alternately, you can pass the base of a file name using out_fnbase. You can specify the output directory with output_dir. The default is to save them in a sub-directory of the images directory called 'map', which will be created if needed.

If combine_feats = TRUE, the geometries from the folders in x will be combined into a single KML or Shapefile. For this to happen, you must also pass a value for combine_fn (a base for the file name of the combined features).

Functions

uas_exp_shp: Export flight info to Shapefile

See also