Create flight metadata text files

Create and/or edit flight metadata text files

uas_metadata_make(
  x,
  md_file = "metadata.txt",
  md_suffix = NULL,
  make_new = FALSE,
  overwrite = FALSE,
  open = TRUE,
  md_flds = uas_getflds(),
  md_template = NULL,
  read_uasinfo = FALSE,
  use_system_editor = FALSE,
  quiet = FALSE
)

Arguments

x

A character vector of directories, or a uas_info object

md_file

Name of the file to create

md_suffix

A suffix to append to the file names

make_new

Create new metadata file(s), Logical

overwrite

Overwrite existing metadata files, Logical

open

Open the file for editing

md_flds

A character vector of field names

md_template

A template metadata file

read_uasinfo

Incorporate existing flight metadata already saved in x (assuming x is a uas_info object)

use_system_editor

Whether to open metadata files in the operating system text editor (as opposed to RStudio), Logical

quiet

Suppress messages

Value

A character vector of the external metadata text file(s) created.

Details

This creates, and/or opens for editing, supplemental metadata text file(s) in image folders.

If make_new = TRUE, metadata files will be created in each directory of x. x can be either a character vector or a uas_info object. The files will be named based on the value of md_file, with an option to append a suffix md_suffix. Both md_file and md_suffix should be of length 1 or equal to the number of x.

If open = TRUE, the text file(s) will be opened. To open existing metadata text files, let open = TRUE and make_new = FALSE. By default, files will be opened with whichever text editor is in use by R (see file.edit). To use the system text editor, let use_system_editor = TRUE.

To customize the fields that are added to flight metadata files, run uas_setflds first. To pre-populate some of the fields, pass the path to an existing metadata file as the value for md_template. md_template can either be a local file or online (example).

See also