Flight summaries

Creates image collection summaries for individual flights (folders)

uas_report(
  x,
  col = NULL,
  group_img = TRUE,
  thumbnails = FALSE,
  show_local_dir = TRUE,
  report_title = "Flight Summary",
  attachments = c("mcp_kml", "ctr_kml")[0],
  output_dir = NULL,
  create_dir = TRUE,
  output_file = NULL,
  overwrite_html = FALSE,
  open_report = FALSE,
  self_contained = TRUE,
  png_map = FALSE,
  png_exp = 0.2,
  overwrite_png = FALSE,
  google_api = NULL,
  report_rmd = NULL,
  header_html = NULL,
  footer_html = NULL,
  use_tmpdir = FALSE,
  quiet = FALSE,
  show_gps_coord = NULL
)

Arguments

x

A list of class 'uas_info'

col

Color value(s) of the centroids and/or footprints

group_img

Group images within ~1m of each other into 1 point

thumbnails

Create thumbails

show_local_dir

Show the local image directory, TF

report_title

Title to appear at the top of the summary

attachments

Supplementary files to create and link to the flight summary, see Details.

output_dir

If NULL, then will be placed in a 'map' sub-directory of the images

create_dir

Create the output directory if it doesn't exist

output_file

Name of the HTML file. If NULL a default based on the name of the input directory is chosen.

overwrite_html

Overwrite existing HTML files without warning, YN

open_report

Open the HTML file in a browser

self_contained

Make the output HTML file self-contained

png_map

Whether to create a PNG version of the map. May be T/F, or dimensions of the output image in pixels (see Details)

png_exp

A proportion to expand the bounding box of the PNG map, see Details.

overwrite_png

Overwrite existing PNG files without warning, YN

google_api

API key for Google Static Maps, see Details.

report_rmd

Rmd template used to generate the HTML file. See Details.

header_html

A HTML file name or URL to use as the header

footer_html

A HTML file name or URL to use as the footer

use_tmpdir

Use the temp dir for processing

quiet

TRUE to supress printing of the pandoc command line

show_gps_coord

Show GPS coordinates of images in the pop-up windows, YN (deprecated)

Value

The HTML file name(s) of the flight summaries generated

Details

This will generate HTML report(s) of the images in the UAS metadata object based.

group_img determinies whether images at the same location are represented by a single point on the map. This is common with multi-spectral sensors that take generate multiple images per location. 'Same location' is determined by looking at the 5th decimal place of the x and y geographic coordinates (~1m),

If no value for output_dir is passed, the report will be saved in a sub-directory of the image directory called 'map'. This sub-directory will be created if create_dir = TRUE.

self_contained determines whether the HTML file(s) created will have all the JavaScript and CSS files embedded in the HTML file itself, or saved in a subdirectory called 'libs'. If saving several reports to one output directory, If saving multiple HTML reports to the same output directory, passing self_contained = FALSE is more efficient

The HTML report is generated from a RMarkdown file. If you know how to edit RMarkdown, you can modify the default template and pass the filename of your preferred template using the report_rmd argument.

png_map controls whether a PNG version of the map will be created in output_dir. If TRUE, a PNG file at the default dimensions (480x480) will be created. If a single integer is passed, it will be taken to be the width and height on the PNG file in pixels. png_exp is a percentage of the points bounding box that will be used as a buffer for the background map. If the map seems too cropped, or you get a warning message about rows removed, try increasing it. By default, the background image will be a satellite photo from Google Maps. However this requires a valid API Key for the Google Maps Static service, which you pass with google_api (for details see https://developers.google.com/maps/documentation/maps-static/. Alternately you can save your API key with ggmap::register_google(), after which it will be read automatically. If a Google API key is not found, a terrain map from Stamen will be substituted.

attachment specifies which supplementary files to create and link to the flight summary. Choices are ctr_kml and mcp_kml for KML versions of the camera locations and MCP (minimum convex polygon around all images). These KML files will be created in the same output directory as the flight summary.

See also