Create image thumbnails

Move UAS images into sub-directories by group

uas_thumbnails_make(
  x,
  flt_idx = NULL,
  output_dir = NULL,
  tb_width = 400,
  overwrite = FALSE,
  use_magick = FALSE,
  stats = FALSE,
  quiet = FALSE
)

Arguments

x

A list of class 'uas_info'

flt_idx

Flight indices in x to process, integer

output_dir

Output directory

tb_width

Thumbnail width

overwrite

Overwrite existing files

use_magick

Use functions from the magick package

stats

Report the amount of time it takes to create each thumbnail, logical

quiet

Suppress messages

Value

A named list (one element for each directory processed) of thumbnail files created in the output directory

Details

This will create thumbnail images for the drone images in x. flt_idx allows you to specify a subset of flights in x to process. The default output folder is a sub-directory of each image folder called map/tb, which will be created if needed. This location can be overridden with output_dir. The dimensions of the thumbnails is determined by tb_width, from which the height is set automatically.

Thumbnail files will be given an 8-character suffix that looks random but is actually generated from the image contents. This is to prevent clashes when thumnbail files from different flights are 'gathered' into a single folder attached to a Table of Contents folder (see uas_toc).

If use_magick = TRUE, it will use resizing functions from the magick package. This is slower than the equivalent functions from the imager package (the default), but may be necessary if you are processing TIFs and don't have ImageMagick installed on your computer (which imager requires to read TIFs).

See also